Add files via upload
Tried to make a button open a set of tabs for a predefined list of urls specified in temp_data.json. I also broke the popup into separate files for layout, style, and scripts. Eventually background.js should keep track of the pages accessed often and open the other popup which will contextually prompt the user to open a tab group. temp_data.json is a placeholder until we work out how we want to utilize chrome.storage.