Add files via upload
I modified the objects we stored on chrome.storage. I also modified the 'numGroups' function so that it returns the number of existing tab groups that the user has. I also tried to implement that feature where the buttons to open specific groups appear from highest usage (top) to lowest usage (bottom), but there's an issue where "freq" increments for only one group. Finally, ignore all the entire function that's commented off. I'm saving it in case I need to use it later on.