Add files via upload
My attempt to implement a testing framework. It is accessed through the options page and testing can be toggled on and off through modifying an if statement in settings.js. I got it to a point where in the console it will say "loading jasmine..." when you click the button. But I'm not sure whether it successfully loads because the "run tests" button does not show anything when I have a spec that should surely fail. (it expects true to be false)